refactored code for readability
separated modal from being written in Home.jsx to its own component as it should be
