+++ client/src/components/index.js
@@ -1,5 +1,6 @@
 import Header from './Header';
 import SearchBar from './SearchBar';
 import SearchResultsList from './SearchResultsList';
+import NewExpenseModal from './NewExpenseModal';
 
-export { Header, SearchBar, SearchResultsList };
+export { Header, SearchBar, SearchResultsList, NewExpenseModal };
